Adapt to db changes

+++ VanKornoCompose/src/main/java/com/vankorno/vankornocompose/_entities/pic/Dao.kt
@@ -3,7 +3,7 @@ package com.vankorno.vankornocompose._entities.pic
 import android.database.sqlite.SQLiteDatabase
 import com.vankorno.vankornodb.get.getObj
 
-fun SQLiteDatabase.getPic(id: Int) = getObj(_TTTPic, PicEntt()) { ID = id }
+fun SQLiteDatabase.getPic(id: Long) = getObj(_TTTPic, PicEntt()) { ID = id }
