+++ VanKornoCompose/src/main/java/com/vankorno/vankornocompose/_entities/pic/old/to100/to10/V2_Pic.kt
@@ -12,7 +12,7 @@ data class V2_Pic(
                                 val usages: Int = 0,
                           val becameUnused: Long = 0L,
                                  val notes: String = "",
-                           override val id: Int = -1,
+                           override val id: Long = -1L,
 ) : OldEntity
 
 
@@ -34,7 +34,7 @@ object _V2_Pic : OldSchemaBundle<V2_Pic>(
             usages = cursor.getInt(idx++),
             becameUnused = cursor.getLong(idx++),
             notes = cursor.getString(idx++),
-            id = cursor.getInt(idx++)
+            id = cursor.getLong(idx++)
         )
     },
